lots of little code refactoring #495 relocated smaller functions and most events divided code into small modules (layout emulator, support popup, version popup)
