+++ src/js/commandline.js
@@ -496,3 +496,8 @@ $(document).keydown((e) => {
     }
   }
 });
+
+$(document).on("click", "#commandLineMobileButton", () => {
+  CommandlineLists.setCurrent(CommandlineLists.defaultCommands);
+  show();
+});
