+++ src/js/exports.js
@@ -3,7 +3,7 @@
 //try to keep this list short because we need to eliminate it eventually
 global.simplePopups = simplePopups;
 global.sendVerificationEmail = Misc.sendVerificationEmail;
-global.getuid = getuid;
+global.getuid = Misc.getuid;
 
 //these exports are just for debugging in the browser
 global.snapshot = DB.getSnapshot;
