+++ src/js/global-dependencies.js
@@ -50,3 +50,6 @@ import * as TestConfig from "./test-config";
 import * as Loader from "./loader";
 import * as MiniResultChart from "./mini-result-chart";
 import * as SignOutButton from "./sign-out-button";
+import * as SupportPopup from "./support-popup";
+import * as VersionPopup from "./version-popup";
+import * as LayoutEmulator from "./layout-emulator";
