+++ src/js/layout-emulator.js
@@ -1,0 +1,105 @@
+import Config from "./config";
+import * as Misc from "./misc";
+import Layouts from "./layouts";
+
+export function updateEvent(event) {
+  function emulatedLayoutShouldShiftKey(event, newKeyPreview) {
+    if (Config.capsLockBackspace) return event.shiftKey;
+    const isCapsLockHeld = event.originalEvent.getModifierState("CapsLock");
+    if (isCapsLockHeld)
+      return Misc.isASCIILetter(newKeyPreview) !== event.shiftKey;
+    return event.shiftKey;
+  }
+
+  function replaceEventKey(event, keyCode) {
+    const newKey = String.fromCharCode(keyCode);
+    event.keyCode = keyCode;
+    event.charCode = keyCode;
+    event.which = keyCode;
+    event.key = newKey;
+    event.code = "Key" + newKey.toUpperCase();
+  }
+
+  let newEvent = event;
+
+  try {
+    if (Config.layout === "default") {
+      //override the caps lock modifier for the default layout if needed
+      if (Config.capsLockBackspace && Misc.isASCIILetter(newEvent.key)) {
+        replaceEventKey(
+          newEvent,
+          newEvent.shiftKey
+            ? newEvent.key.toUpperCase().charCodeAt(0)
+            : newEvent.key.toLowerCase().charCodeAt(0)
+        );
+      }
+      return newEvent;
+    }
+    const keyEventCodes = [
+      "Backquote",
+      "Digit1",
+      "Digit2",
+      "Digit3",
+      "Digit4",
+      "Digit5",
+      "Digit6",
+      "Digit7",
+      "Digit8",
+      "Digit9",
+      "Digit0",
+      "Minus",
+      "Equal",
+      "KeyQ",
+      "KeyW",
+      "KeyE",
+      "KeyR",
+      "KeyT",
+      "KeyY",
+      "KeyU",
+      "KeyI",
+      "KeyO",
+      "KeyP",
+      "BracketLeft",
+      "BracketRight",
+      "Backslash",
+      "KeyA",
+      "KeyS",
+      "KeyD",
+      "KeyF",
+      "KeyG",
+      "KeyH",
+      "KeyJ",
+      "KeyK",
+      "KeyL",
+      "Semicolon",
+      "Quote",
+      "IntlBackslash",
+      "KeyZ",
+      "KeyX",
+      "KeyC",
+      "KeyV",
+      "KeyB",
+      "KeyN",
+      "KeyM",
+      "Comma",
+      "Period",
+      "Slash",
+      "Space",
+    ];
+    const layoutMap = Layouts[Config.layout].keys;
+
+    let mapIndex;
+    for (let i = 0; i < keyEventCodes.length; i++) {
+      if (newEvent.code == keyEventCodes[i]) {
+        mapIndex = i;
+      }
+    }
+    const newKeyPreview = layoutMap[mapIndex][0];
+    const shift = emulatedLayoutShouldShiftKey(newEvent, newKeyPreview) ? 1 : 0;
+    const newKey = layoutMap[mapIndex][shift];
+    replaceEventKey(newEvent, newKey.charCodeAt(0));
+  } catch (e) {
+    return event;
+  }
+  return newEvent;
+}
