+++ src/js/test/focus.js
@@ -19,3 +19,12 @@ export function set(foc) {
     $("#middle").removeClass("focus");
   }
 }
+
+$(document).mousemove(function (event) {
+  if (
+    $("#top").hasClass("focus") &&
+    (event.originalEvent.movementX > 0 || event.originalEvent.movementY > 0)
+  ) {
+    set(false);
+  }
+});
