+++ src/js/version-popup.js
@@ -1,0 +1,14 @@
+$(document.body).on("click", ".version", () => {
+  $("#versionHistoryWrapper")
+    .css("opacity", 0)
+    .removeClass("hidden")
+    .animate({ opacity: 1 }, 125);
+});
+
+$(document.body).on("click", "#versionHistoryWrapper", () => {
+  $("#versionHistoryWrapper")
+    .css("opacity", 1)
+    .animate({ opacity: 0 }, 125, () => {
+      $("#versionHistoryWrapper").addClass("hidden");
+    });
+});
